fix(lint): re-enable bugprone-narrowing-conversions
All eight findings were an unsigned container size reaching a signed or
floating-point parameter without saying so: `non_seq.size() *
destroy_fraction_` and `model.num_vars() * destroy_fraction_` in the LNS
destroy step, the four `rng.integers(0, v.size())` bounds in
`set_moves`, and two in the perturb tests. Each is now an explicit
static_cast to the destination type; no arithmetic changed, so the
search trajectory is untouched and the suite is unchanged at 306/306
fast.

static_cast is used rather than a C-style cast because
modernize-avoid-c-style-cast is still in the ratchet and would have
flagged the alternative.

Ratchet: 13 checks / 262 findings -> 12 / 254, re-derived by deleting
the remaining lines from .clang-tidy and re-sweeping, not by subtracting.

+++ src/lns.cpp
@@ -82,13 +82,15 @@ bool LNS::destroy_repair(Model& model, ViolationManager& vm, RNG& rng, double re
                 non_seq.push_back(i);
             }
         }
-        int n_non_seq_destroy = std::max(0, static_cast<int>(non_seq.size() * destroy_fraction_));
+        int n_non_seq_destroy =
+            std::max(0, static_cast<int>(static_cast<double>(non_seq.size()) * destroy_fraction_));
         rng.shuffle(non_seq);
         non_seq.resize(n_non_seq_destroy);
         var_indices.insert(var_indices.end(), non_seq.begin(), non_seq.end());
     } else {
         // Uniform random destroy (no sequences registered)
-        int n_destroy = std::max(1, static_cast<int>(model.num_vars() * destroy_fraction_));
+        int n_destroy = std::max(
+            1, static_cast<int>(static_cast<double>(model.num_vars()) * destroy_fraction_));
         var_indices.resize(model.num_vars());
         std::iota(var_indices.begin(), var_indices.end(), 0);
         rng.shuffle(var_indices);
